Refactor the build graph
This implements our own graph type such that the graph can own the nodes
it holds. This greatly simplifies how the build graph is interacted with
and also reduces the memory footprint of the program.

Note: petgraph does not support this, so our own graph had to be
implemented instead. There is one significant limitation that we inherit
from `indexmap`. Nodes can never be removed from the graph, only added.
Currently, this doesn't matter very much because we only ever add nodes
to the graph. However, in the future when implicit dependencies are
implemented we will need to be able to remove nodes.